Updated routing and NavBar code
