--- swin_rover_app/src/views/projects/projects.css
@@ -1,121 +1,0 @@
-/* ********** Home Page Styles ********** */
-
-.title {
-    font-size: var(--title-size);
-    line-height: 1;
-    width: 90%;
-}
-.headline {
-    font-size: var(--headline-size);
-}
-.medium-text {
-    font-size: var(--medium-size);
-}
-.small-text {
-    font-size: var(--small-size);
-}
-.top-element {
-    margin-top: 17.5vh;
-}
-.width-100 {
-    width: 100%;
-}
-.relative {
-    position: relative;
-    height: 45vh;
-}
-.centre-content {
-    height: 100%;
-    display: flex;
-    align-items: end;
-    justify-content: center;
-}
-.social-links-box {
-    height: 40vh;
-    width: 100%;
-}
-.social-links-box-inner {
-    width: 100%;
-    border: 1px solid #303030;
-    padding: 5vh 0;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-.social-links-box-inner.left {
-    border-radius: 3em 0 0 3em;
-}
-.social-links-box-inner.right {
-    border-radius: 0 3em 3em 0;
-}
-
-.featured-project {
-    background:var(--red-color);
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    height:40vh;
-    border-radius:1vw;
-}
-
-.featured-project .tag {
-    padding:2vh 1.5vw;
-    margin:0;
-}
-
-.img-container img {
-    width:100%;
-    height:40vh;
-    border-radius: 1vw;
-    object-fit: cover;
-}
-
-.featured-project .project-details {
-    border-top: 2px solid rgba(255,255,255,0.2);
-    padding:3vh 1.5vw;
-    display: flex;
-    justify-content: space-between;
-    align-items: flex-end;
-}
-
-.featured-project h2 {
-    font-size:var(--medium-size);
-    font-weight:400;
-    width:50%;
-}
-
-.featured-project .tag span {
-    color:white;
-}
-
-@media screen and (max-width:900px) {
-    .chunk > div:nth-child(2), .chunk > div:nth-child(3), .featured-project, .chunk img {
-        height:25vh;
-        border-radius: 3vw;
-    }
-
-    .chunk-no-border {
-        margin-top:12.5vh;
-    }
-
-    .title {
-        margin-bottom:0;
-        width:100%;
-    }
-
-    .featured-project .tag {
-        padding: 0.75vh 2vw;
-    }
-
-    .featured-project .project-details {
-        padding: 0.75vh 2vw;
-    }
-
-    .featured-project h2 {
-        width:60%;
-    }
-
-    .container img {
-        height:30vh;
-    }
-}
